Separate out skipping index ast builder
